% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_loss_given_depth_change.R
\name{area_loss_given_depth_change}
\alias{area_loss_given_depth_change}
\title{Calculate percent loss of lake area with Xft change in lake depth}
\usage{
area_loss_given_depth_change(df, depth_change_ft = 1)
}
\arguments{
\item{df}{data frame with information about "WBIC", "lake", depth_feet",
and "area_acres". Input data frame may have other columns, but
must have these.}

\item{depth_change_ft}{Specify what reduction in lake depth from maximum
lake depth to evaluate. Defaults to 1 (ft).}
}
\value{
data frame with the following columns:
\item{WBIC}{Wisconsin Water Body Identification Code (WBIC) of lake}
\item{lake}{name of lake as included in the input data frame. Typically
            something like "easthorsehead", just to help with quick
            identification)}
\item{max_depth}{maximum lake depth (ft)}
\item{max_area}{lake area corresponding to macimum lake depth (acres)}
\item{lower_depth}{lake depth assuming Xft reduction (as input, default is
                   1ft reduction) (ft)}
\item{lower_area}{volume corresponding to lowered depth (acres)}
\item{area_pcnt_loss}{percentage change in lake area relative to maximum
                     lake area given a Xft reduction in lake depth (as
                     input, default is 1ft reduction)}
}
\description{
Given a data frame with information about "WBIC", "depth_feet", and
"area_acres" and what reduction in max depth (ft) to assume, calculates
the estimated change in lake area (%) from the maximum lake depth/area
Assumes a linear change in depth and area between specified contour
intervals in order to approximate lower lake area
}
