% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_volumes.R
\name{calculate_volumes}
\alias{calculate_volumes}
\title{Calculate lake volumes (acre-ft) from lake areas (acres) and lake depths (ft)}
\usage{
calculate_volumes(df, method = "trapezoidal")
}
\arguments{
\item{df}{data frame with information about "WBIC", "depth_feet", and
"area_acres". Input data frame may have other columns, but must
have these.}

\item{method}{Specify which volume estimation approach to use. Options
include "trapezoidal" or "conical". Defaults to "trapezoidal".}
}
\value{
the input data frame with an additional column for "volume_acre_ft"
        appended, which represents the lake volume at the given lake depth
        (acre-ft).
}
\description{
Given a data frame with information about "WBIC", "depth_feet", and
"area_acres" and which type of volume estimate to use ("trapezoidal" or
"conical"), calculates the lake volume in acre-ft associated with each lake
depth. Note that this approach assumes that information about "depth_feet" is
arranged such that the maximum depth value corresponds with the lake surface
and the maximum lake volume value, while a depth value of 0 corresponds with
the lake bottom and a lake volume of 0.
}
