% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_thief_volumes.R
\name{convert_data_thief_volumes}
\alias{convert_data_thief_volumes}
\title{Convert depth vs. volume information obtained via DataThief into standard format}
\usage{
convert_data_thief_volumes(path, max_file, flip = TRUE, depth_interval = 3)
}
\arguments{
\item{path}{Path of directory containing all txt files with depth and
volume info (extracted from bathymetry maps using
DataThief), e.g. "data-raw/CS_DataThief". See details in
description about expected format of these txt files.}

\item{max_file}{Name of csv file with information about maximum areas of
each lake, e.g. "WI_maxes.csv". See details in description
about expected format of this csv file.}

\item{flip}{Indicates whether should flip depths so that the maximum
depth/volume corresponds to the lake surface and a depth/volume
of zero corresponds to the lake bottom (TRUE) or keep as-is, with
a depth of zero corresponding to the lake surface (FALSE).
Defaults to TRUE to flip.}

\item{depth_interval}{optional parameter specifies what depth contour
interval to extrapolate volume information at. Defaults
to 3ft.}
}
\value{
a data frame with the following columns:
\item{WBIC}{Wisconsin Water Body Identification Code (WBIC) of lake}
\item{lake}{name of lake as included in csv filename (will be a single
            word, e.g. "easthorsehead", just to help with quick
            identification)}
\item{depth_feet}{lake depth (ft)}
\item{volume_acre_ft}{lake volume at this lake depth (acre-ft)}
}
\description{
Given the path of a directory where txt files from DataThief output live,
reads in information about depth (ft) vs. volume (acre-ft) relationship and
converts to standard format for other analyses.
}
\details{
Expects txt files with depth vs. volume information to adhere to the
following conventions:
* All txt files are stored in the same directory (specified by "path")
* Filenames are "singlelakename_WBIC.csv", e.g. "easthorsehead_1523000.txt"
* Each file starts with a header from DataThief that should be deleted
* Columns are unnamed but represent (in order): "volume_acre_ft", "depth_feet"
* A depth/volume value of 0 represents the top of the lake, maximum values
  represent the bottom of hte lake.
* Every txt file following this convention should have a corresponding entry
  in the maximums txt file.

Expects that information about maximum lake depth (ft) and lake volume
(acre-ft) adheres to the following conventions:
 * Information is in a single csv file (specified by "max_file") stored
   in the same directory as includes the individual txt files.
 * Every lake with a DataThief txt file should have an entry here
 * Columns include: "WBIC", "Depth_feet", and "Volume_acre_ft" (may include
   others as well, but unused here).
 * Missing values are denoted by NA (capitalized).
 * If there is a value for "Depth_feet", but "Volume_acre_ft" is NA,
   will use maximum depth plus the DataThief information to estimate what
   maximum volume is likely to be (and vice versa, if has maximum volume but
   maximum depth is NA).

Note that the returned data frame lists the lake volume (acre-ft)
corresponding to lake depths such that 0ft is the lake bottom, 5ft is 5ft
above the the lake bottom, and the maximum depth value represents the surface
of the lake. If the parameter "flip" is set to FALSE, the returned data frame
instead retains the conventions in the DataThief txt files, where 0 values
for depth and volume represent the lake surface and maximum values represent
the bottom of the lake.
}
