% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_proportion_areas.R
\name{convert_proportion_areas}
\alias{convert_proportion_areas}
\title{Convert proportional area hypsography to areas in acres}
\usage{
convert_proportion_areas(
  path = "data-raw/Wisconsin_Hypsography",
  max_area_file = "WI_areas.csv",
  flip = TRUE
)
}
\arguments{
\item{path}{Path of directory containing all csv files with depth and
proportional area info (extracted from bathymetry maps using
imageJ). Defaults to "data-raw/Wisconsin_Hypsography". See
details in description about expected format of these csv files.}

\item{max_area_file}{Name of csv file with information about maximum areas of
each lake. Defaults to "WI_areas.csv". See details in
description about expected format of this csv file.}

\item{flip}{Indicates whether should flip depths so that the maximum depth
corresponds to the lake surface and a depth of zero corresponds
to the lake bottom (TRUE) or keep as-is, with a depth of zero
corresponding to the lake surface (FALSE). Defaults to TRUE to
flip.}
}
\value{
a data frame with the following columns:
\item{WBIC}{Wisconsin Water Body Identification Code (WBIC) of lake}
\item{lakename}{name of lake as included in csv filename (will be a single
                word, e.g. "easthorsehead", just to help with quick
                identification)}
\item{depth_ft}{lake depth (ft)}
\item{area_acres}{lake area at this lake depth (acres)}
}
\description{
Given the path of a directory where csv files live, reads in information
about depth (ft) vs. proportional area relationship, maximum areas, and
converts to depth (ft) vs. area (acres) relationships.
}
\details{
Expects csv files with depth vs. proportional area information to adhere to
the following conventions:
* All csv files are stored in the same directory (specified by "path")
* Filenames are "singlelakename_WBIC.csv", e.g. "easthorsehead_1523000.csv"
* Columns are (in order): "depth_ft", "proportion_area"
* A proportion area of "1" corresponds with a depth_ft of "0" and represents
 the maximum area as measured at the top of the lake.
* Every csv file following this convention should have a corresponding entry
  in the maximum lake area csv file.

Expects that information about maximum lake areas (in acres) adheres to the
following conventions:
 * Information is in a single csv file (specified by "max_area_file") stored
   in the same directory as includes the individual csv files.
 * Columns are (in order): "WBIC" and "Area_acres"
 * Every lake with a proportional area csv file should have an entry here

Note that the returned data frame lists the lake area (acres) corresponding
to lake depths such that 0ft is the lake bottom, 5ft is 5ft above the the
lake bottom, and the maximum depth value represents the surface of the lake.
If the parameter "flip" is set to FALSE, the returned data frame instead
retains the conventions in the proportional area csv files, where lake areas
correspond to lake depth contours such that 0ft is the lake surface, 5ft is
5ft below the lake surface, and the maximum depth value represents the bottom
of the lake.
}
