% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_change_given_area_loss.R
\name{depth_change_given_area_loss}
\alias{depth_change_given_area_loss}
\title{Calculate change in lake depth with X percent loss of lake area}
\usage{
depth_change_given_area_loss(df, pcnt_loss = 10)
}
\arguments{
\item{df}{data frame with information about "WBIC", "lakename", depth_feet",
and "area_acres". Input data frame may have other columns, but
must have these.}

\item{pcnt_loss}{Specify what percentage loss of lake area from maximum
lake area to evaluate. Defaults to 10 (percent).}
}
\value{
data frame with the following columns:
\item{WBIC}{Wisconsin Water Body Identification Code (WBIC) of lake}
\item{lakename}{name of lake as included in the input data frame. Typically
                something like "easthorsehead", just to help with quick
                identification)}
\item{max_area}{maximum lake area (acres)}
\item{max_depth}{lake depth corresponding to maximum lake area (ft)}
\item{lower_area}{area assuming X pcnt loss (as input, default is 10 pcnt
                 loss) (acres)}
\item{lower_depth}{lake depth corresponding to lowered lake area (ft)}
\item{depth_change}{change in lake depth corresponding to X pcnt loss (as
                    input, default is 10 pcnt loss) in lake area (ft)}
}
\description{
Given a data frame with information about "WBIC", "depth_feet", and
"area_acres" and what percentage area loss to assume, calculates the
estimated change in lake depth from the maximum lake depth/area Assumes a
linear change in depth and area between specified contour intervals in
order to approximate lower lake depth.
}
