% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_loss_given_depth_change.R
\name{volume_loss_given_depth_change}
\alias{volume_loss_given_depth_change}
\title{Calculate percent loss of lake volume with Xft change in lake depth}
\usage{
volume_loss_given_depth_change(df, depth_change_ft = 1)
}
\arguments{
\item{df}{data frame with information about "WBIC", "lake", depth_feet",
and "volume_acre_ft". Input data frame may have other columns, but
must have these.}

\item{depth_change_ft}{Specify what reduction in lake depth from maximum
lake depth to evaluate. Defaults to 1 (ft).}
}
\value{
data frame with the following columns:
\item{WBIC}{Wisconsin Water Body Identification Code (WBIC) of lake}
\item{lake}{name of lake as included in the input data frame. Typically
            something like "easthorsehead", just to help with quick
            identification)}
\item{max_depth}{maximum lake depth (ft)}
\item{max_vol}{lake volume corresponding to macimum lake depth (acre-ft)}
\item{lower_depth}{lake depth assuming Xft reduction (as input, default is
                   1ft reduction) (ft)}
\item{lower_vol}{volume corresponding to lowered depth (acre-ft)}
\item{vol_pcnt_loss}{percentage change in lake volume relative to maximum
                     lake volume given a Xft reduction in lake depth (as
                     input, default is 1ft reduction)}
}
\description{
Given a data frame with information about "WBIC", "depth_feet", and
"volume_acre_ft" and what reduction in max depth (ft) to assume, calculates
the estimated change in lake volume (%) from the maximum lake depth/volume.
Assumes a linear change in depth and volume between specified contour
intervals in order to approximate lower lake volume.
}
